Remove all of the pre-0.97.0.0 deprecations.
Remove the 0.97.0.10 plot group deprecations that were never used.
