+++ src/com/palmergames/bukkit/towny/TownyFormatter.java
@@ -1117,93 +1117,5 @@ public class TownyFormatter {
 
 		return webUrl;
 	}
-	
-	/*
-	 * Deprecated methods.
-	 */
-	
-	/**
-	 * @param obj The {@link TownyObject} to get the formatted name from.
-	 * 
-	 * @return The formatted name of the object.
-	 * 
-	 * @deprecated Since 0.96.0.0 use {@link TownyObject#getFormattedName()} instead.
-	 */
-	@Deprecated
-	public static String getFormattedName(TownyObject obj) {
-		return obj.getFormattedName();
-	}
-
-	/**
-	 * @param resident The {@link Resident} to get the formatted name from.
-	 *                    
-	 * @return The formatted name of the object.
-	 * 
-	 * @deprecated Since 0.96.0.0 use {@link Resident#getFormattedName()} instead.
-	 */
-	@Deprecated
-	public static String getFormattedResidentName(Resident resident) {
-		return resident.getFormattedName();
-	}
-
-	/**
-	 * @param town The {@link Town} to get the formatted name from.
-	 *                
-	 * @return The formatted name of the object.
-	 * 
-	 * @deprecated Since 0.96.0.0 use {@link Town#getFormattedName()} instead.
-	 */
-	@Deprecated
-	public static String getFormattedTownName(Town town) {
-		return town.getFormattedName();
-	}
-
-	/**
-	 * @param nation The {@link Nation} to get the formatted name from.
-	 * 
-	 * @return The formatted name of the object.
-	 *
-	 * @deprecated Since 0.96.0.0 use {@link Nation#getFormattedName()} instead.
-	 */
-	@Deprecated
-	public static String getFormattedNationName(Nation nation) {
-		return nation.getFormattedName();
-	}
-
-	/**
-	 * @param resident The {@link Resident} to get the formatted title name from.
-	 *                    
-	 * @return The formatted title name of the resident.
-	 * 
-	 * @deprecated Since 0.96.0.0 use {@link Resident#getFormattedTitleName()} instead.
-	 */
-	@Deprecated
-	public static String getFormattedResidentTitleName(Resident resident) {
-		return resident.getFormattedTitleName();
-	}
-	
-	/**
-	 * @param resident The {@link Resident} to get the king or mayor prefix from.
-	 *                    
-	 * @return The king or mayor prefix of the resident.
-	 *
-	 * @deprecated Since 0.96.0.0 use {@link Resident#getNamePrefix()} instead.
-	 */
-	@Deprecated	
-	public static String getNamePrefix(Resident resident) {
-		return resident.getNamePrefix();	
-	}	
-
-	/**
-	 * @param resident The {@link Resident} to get the king or mayor postfix from.
-	 *                    
-	 * @return The king or mayor postfix of the resident.
-	 *
-	 * @deprecated Since 0.96.0.0 use {@link Resident#getNamePostfix()} instead.
-	 */
-	@Deprecated	
-	public static String getNamePostfix(Resident resident) {
-		return resident.getNamePostfix();
-	}	
 
 }
