+++ src/com/palmergames/bukkit/towny/db/TownyDataSource.java
@@ -6,6 +6,7 @@ import com.palmergames.bukkit.towny.TownyMessaging;
 import com.palmergames.bukkit.towny.TownySettings;
 import com.palmergames.bukkit.towny.TownyUniverse;
 import com.palmergames.bukkit.towny.exceptions.AlreadyRegisteredException;
+import com.palmergames.bukkit.towny.exceptions.InvalidNameException;
 import com.palmergames.bukkit.towny.exceptions.NotRegisteredException;
 import com.palmergames.bukkit.towny.object.Nation;
 import com.palmergames.bukkit.towny.object.PlotGroup;
@@ -18,9 +19,9 @@ import com.palmergames.bukkit.towny.regen.PlotBlockData;
 import org.jetbrains.annotations.Nullable;
 
 import java.io.IOException;
+import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.Set;
 import java.util.UUID;
 import java.util.concurrent.locks.Lock;
 import java.util.concurrent.locks.ReentrantLock;
@@ -314,7 +315,9 @@ public abstract class TownyDataSource {
 	 * @return list of all towny residents
 	 */
 	@Deprecated
-	abstract public List<Resident> getResidents();
+	public List<Resident> getResidents() {
+		return new ArrayList<>(TownyUniverse.getInstance().getResidents());
+	}
 	
 	/**
 	 * @deprecated since 0.97.5.18 use {@link TownyUniverse#getGroups()} instead.
@@ -341,24 +344,7 @@ public abstract class TownyDataSource {
 	@Deprecated
 	abstract public List<Resident> getResidents(UUID[] uuids);
 
-	/**
-	 * @deprecated as of 0.96.6.0. Use {@link TownyUniverse#getResident(String)} instead.
-	 * @param name The name of the resident.
-	 * @return Resident with the given name.
-	 * @throws NotRegisteredException if the Resident does not exist.
-	 */
-	@Deprecated
-	abstract public Resident getResident(String name) throws NotRegisteredException;
-
 	abstract public void removeNation(Nation nation);
-
-	/**
-	 * @deprecated as of 0.96.6.0. Use {@link TownyUniverse#hasResident(String)} instead.
-	 * @param name The name of the resident.
-	 * @return true if the resident exists.
-	 */
-	@Deprecated
-	abstract public boolean hasResident(String name);
 
 	/**
 	 * @deprecated as of 0.97.5.3, use {@link TownyUniverse#hasTown(String)} instead.
@@ -369,7 +355,9 @@ public abstract class TownyDataSource {
 	 * @return whether the town exists.
 	 */
 	@Deprecated
-	abstract public boolean hasTown(String name);
+	public boolean hasTown(String name) {
+		return universe.hasTown(name);
+	}
 
 	/**
 	 * @deprecated as of 0.97.5.3, use {@link TownyUniverse#hasNation(String)} instead.
@@ -380,7 +368,9 @@ public abstract class TownyDataSource {
 	 * @return whether the nation with the given name exists.
 	 */
 	@Deprecated
-	abstract public boolean hasNation(String name);
+	public boolean hasNation(String name) {
+		return universe.hasNation(name);
+	}
 
 	/**
 	 * @deprecated as of 0.97.5.18, use {@link TownyAPI#getTowns(String[])} instead.
@@ -398,7 +388,9 @@ public abstract class TownyDataSource {
 	 * @return a list of all towns.
 	 */
 	@Deprecated
-	abstract public List<Town> getTowns();
+	public List<Town> getTowns() {
+		return new ArrayList<>(TownyUniverse.getInstance().getTowns());
+	}
 
 	/**
 	 * @deprecated as of 0.97.5.3, Use {@link TownyUniverse#getTown(String)} instead.
@@ -409,7 +401,14 @@ public abstract class TownyDataSource {
 	 * @throws NotRegisteredException Town does not exist.
 	 */
 	@Deprecated
-	abstract public Town getTown(String name) throws NotRegisteredException;
+	public Town getTown(String name) throws NotRegisteredException {
+		Town town = universe.getTown(name);
+		
+		if (town == null)
+			throw new NotRegisteredException(String.format("The town with name '%s' is not registered!", name));
+		
+		return town;
+	}
 
 	/**
 	 * @deprecated as of 0.97.5.3, Use {@link TownyUniverse#getTown(UUID)} instead.
@@ -423,7 +422,14 @@ public abstract class TownyDataSource {
 	 * @throws NotRegisteredException Thrown if town doesn't exist.
 	 */
 	@Deprecated
-	abstract public Town getTown(UUID uuid) throws NotRegisteredException;
+	public Town getTown(UUID uuid) throws NotRegisteredException {
+		Town town = universe.getTown(uuid);	
+		
+		if (town == null)
+			throw new NotRegisteredException(String.format("The town with uuid '%s' is not registered.", uuid));
+		
+		return town;
+	}
 
 	/**
 	 * @deprecated as of 0.97.5.18, use {@link TownyAPI#getNations(String[])} instead.
@@ -439,7 +445,9 @@ public abstract class TownyDataSource {
 	 * @return all nations.
 	 */
 	@Deprecated
-	abstract public List<Nation> getNations();
+	public List<Nation> getNations() {
+		return new ArrayList<>(universe.getNations());
+	}
 
 	/**
 	 * @deprecated as of 0.97.5.3, Please use {@link TownyUniverse#getNation(String)} instead.
@@ -451,7 +459,14 @@ public abstract class TownyDataSource {
 	 * @throws NotRegisteredException if no nation is found matching the given name.
 	 */
 	@Deprecated
-	abstract public Nation getNation(String name) throws NotRegisteredException;
+	public Nation getNation(String name) throws NotRegisteredException {
+		Nation nation = universe.getNation(name);
+
+		if (nation == null)
+			throw new NotRegisteredException(String.format("The nation '%s' is not registered.", name));
+
+		return nation;
+	}
 
 	/**
 	 * @deprecated as of 0.97.5.3, Use {@link TownyUniverse#getNation(UUID)} instead.
@@ -463,7 +478,14 @@ public abstract class TownyDataSource {
 	 * @throws NotRegisteredException if no nation is found matching the given UUID.
 	 */
 	@Deprecated
-	abstract public Nation getNation(UUID uuid) throws NotRegisteredException;
+	public Nation getNation(UUID uuid) throws NotRegisteredException {
+		Nation nation = universe.getNation(uuid);
+		
+		if (nation == null)
+			throw new NotRegisteredException(String.format("The nation with uuid '%s' is not registered.", uuid.toString()));
+		
+		return nation;
+	}
 
 	/**
 	 * @deprecated as of 0.97.5.18, Use {@link TownyUniverse#getWorld(String)} instead.
@@ -482,27 +504,6 @@ public abstract class TownyDataSource {
 	 */
 	@Deprecated
 	abstract public List<TownyWorld> getWorlds();
-
-	/**
-	 * @deprecated as of 0.96.3.0, Use {@link Town#getHomeblockWorld()} instead.
-	 * 
-	 * Legacy method to get a world associated with a town.
-	 * 
-	 * @param townName The name of a town.
-	 * 
-	 * @return Returns a {@link TownyWorld} associated with the town.
-	 */
-	@Deprecated // TODO: Scrap worlds holding Towns. Towns' homeblocks should be reliable enough to return a world when needed (if we need it at all anymore.)
-	public TownyWorld getTownWorld(String townName) {
-
-		for (TownyWorld world : universe.getTownyWorlds()) {
-			if (world.hasTown(townName))
-				return world;
-		}
-
-		// If this has failed the Town has no land claimed at all but should be given a world regardless.
-		return universe.getTownyWorlds().get(0);
-	}
 
 	abstract public void removeResident(Resident resident);
 
@@ -531,7 +532,13 @@ public abstract class TownyDataSource {
 	 * @throws NotRegisteredException thrown if town has an invalid name.
 	 */
 	@Deprecated
-	abstract public void newTown(String name) throws AlreadyRegisteredException, NotRegisteredException;
+	public void newTown(String name) throws AlreadyRegisteredException, NotRegisteredException {
+		try {
+			universe.newTown(name);
+		} catch (InvalidNameException e) {
+			throw new NotRegisteredException(e.getMessage());
+		}
+	}
 
 	abstract public void newNation(String name) throws AlreadyRegisteredException, NotRegisteredException;
 
@@ -548,30 +555,6 @@ public abstract class TownyDataSource {
 	abstract public void removeJail(Jail jail);
 	
 	abstract public void removePlotGroup(PlotGroup group);
-
-	/**
-	 * @deprecated as of 0.96.4.0, We do not advise messing with the Residents Map.
-	 * 
-	 * @return Returns a {@link Set} of the Residents Map
-	 */
-	@Deprecated
-	abstract public Set<String> getResidentKeys();
-
-	/**
-	 * @deprecated as of 0.96.4.0, We do not advise messing with the Towns Map.
-	 * 
-	 * @return Returns a {@link Set} of the Towns Map
-	 */
-	@Deprecated
-	abstract public Set<String> getTownsKeys();
-
-	/**
-	 * @deprecated as of 0.96.4.0, We do not advise messing with the Nations Map.
-	 * 
-	 * @return Returns a {@link Set} of the Nations Map
-	 */
-	@Deprecated
-	abstract public Set<String> getNationsKeys();
 
 	/**
 	 * @deprecated as of 0.97.5.18 use {@link TownyAPI#getTownsWithoutNation} instead.
