+++ src/com/palmergames/bukkit/towny/object/Town.java
@@ -27,7 +27,6 @@ import com.palmergames.bukkit.towny.object.metadata.CustomDataField;
 import com.palmergames.bukkit.towny.permissions.TownyPerms;
 import com.palmergames.bukkit.util.BukkitTools;
 import com.palmergames.util.MathUtil;
-import com.palmergames.util.StringMgmt;
 import net.kyori.adventure.audience.Audience;
 import org.bukkit.Bukkit;
 import org.bukkit.Location;
@@ -284,16 +283,6 @@ public class Town extends Government implements TownBlockOwner {
 		return Collections.unmodifiableList(residents);
 	}
 
-	/**
-	 * @deprecated as of 0.96.3.0, use {@link #getRank(String)} instead.
-	 * @return Returns a list of residents with the assistant rank.
-	 */
-	@Deprecated
-	public List<Resident> getAssistants() {
-
-	    return getRank("assistant");
-	}
-
 	public List<Resident> getRank(String rank) {
 
 		List<Resident> residentsWithRank = new ArrayList<>();
@@ -328,19 +317,6 @@ public class Town extends Government implements TownBlockOwner {
 		
 		Resident resident = TownyAPI.getInstance().getResident(uuid);
 		return resident != null && hasResident(resident);
-	}
-	
-	/**
-	 * @deprecated Since 0.96.3.0, use {@link Resident#hasTownRank(String)} (using "assistant" as argument) instead.
-	 * Whether a resident has an assistant role or not.
-	 * 
-	 * @param resident - Resident to check for a rank.
-	 * @return A true if the resident is an assistant, false otherwise.
-	 */
-	@Deprecated
-	public boolean hasAssistant(Resident resident) {
-
-		return resident.hasTownRank("assistant");
 	}
 	
 	public boolean hasResidentWithRank(Resident resident, String rank) {
@@ -1274,22 +1250,6 @@ public class Town extends Government implements TownBlockOwner {
 		
 		return Collections.unmodifiableCollection(plotGroups.values());
 	}
-
-	// Method is inefficient compared to getting the group from name.
-	/**
-	 * @deprecated since 0.97.0.11 for being unused.
-	 */
-	@Deprecated
-	public PlotGroup getObjectGroupFromID(UUID plotGroupId) {
-		if (hasPlotGroups()) {
-			for (PlotGroup pg : getPlotGroups()) {
-				if (pg.getID().equals(plotGroupId)) 
-					return pg;
-			}
-		}
-		
-		return null;
-	}
 	
 	public boolean hasPlotGroups() {
 		return plotGroups != null;
@@ -1426,49 +1386,6 @@ public class Town extends Government implements TownBlockOwner {
 	
 	public boolean hasNationZoneOverride() {
 		return nationZoneOverride > 0;
-	}
-	
-	/**
-	 * @deprecated As of 0.96.0.0+ please use {@link EconomyAccount#getWorld()} instead.
-	 * 
-	 * @return The world this resides in.
-	 */
-	@Deprecated
-	public World getBukkitWorld() {
-		if (hasWorld()) {
-			return BukkitTools.getWorld(getHomeblockWorld().getName());
-		} else {
-			return BukkitTools.getWorlds().get(0);
-		}
-	}
-
-	/**
-	 * @deprecated As of 0.96.0.0+ please use {@link EconomyAccount#getName()} instead.
-	 * 
-	 * @return The name of the economy account.
-	 */
-	@Deprecated
-	public String getEconomyName() {
-		return StringMgmt.trimMaxLength(Town.ECONOMY_ACCOUNT_PREFIX + getName(), 32);
-	}
-	
-	/**
-	 * @deprecated As of 0.96.2.0, please use {@link #getBoard()} instead.
-	 * 
-	 * @return getBoard()
-	 */
-	@Deprecated
-	public String getTownBoard() {
-		return getBoard();
-	}
-
-	/**
-	 * @return The UUID for this town.
-	 * @deprecated as of 0.96.4.0, use {@link #getUUID()} instead.
-	 */
-	@Deprecated
-	public UUID getUuid() {
-		return getUUID();
 	}
 
 	public long getJoinedNationAt() {
