+++ src/com/palmergames/bukkit/towny/object/economy/Account.java
@@ -279,32 +279,6 @@ public abstract class Account implements Nameable {
 		addObserver(auditor);
 	}
 	
-	// Legacy Compatibility Methods.
-	
-	/**
-	 * @deprecated As of 0.96.1.11, use {@link #deposit(double, String)} instead.
-	 * 
-	 * @param amount The amount to add.
-	 * @param reason The reason for adding.
-	 * @return boolean indicating success.
-	 */
-	@Deprecated
-	public boolean collect(double amount, String reason) {
-		return deposit(amount, reason);
-	}
-	
-	/**
-	 * @deprecated As of 0.96.1.11, use {@link #withdraw(double, String)} instead.
-	 *
-	 * @param amount The amount to subtract.
-	 * @param reason The reason for subcracting.
-	 * @return boolean indicating success.
-	 */
-	@Deprecated
-	public boolean pay(double amount, String reason) {
-		return withdraw(amount, reason);
-	}
-
 	class CachedBalance {
 		private double balance = 0;
 		private long time;
