+++ src/com/palmergames/bukkit/towny/utils/CombatUtil.java
@@ -371,20 +371,6 @@ public class CombatUtil {
 	public static boolean isWorldPvP(TownyWorld world) {
 		return (world.isForcePVP() || world.isPVP());
 	}
-
-	/**
-	 * @deprecated as of 0.96.2.20 use {@link CombatUtil#preventFriendlyFire(Player, Player, TownyWorld) instead}
-	 * Should we be preventing friendly fire?
-	 * 
-	 * @param attacker - Attacking Player
-	 * @param defender - Defending Player (receiving damage)
-	 * 
-	 * @return true if we should cancel damage.
-	 */
-	@Deprecated
-	public static boolean preventFriendlyFire(Player attacker, Player defender) {
-		return preventFriendlyFire(attacker, defender, TownyAPI.getInstance().getTownyWorld(attacker.getWorld().getName()));
-	}
 	
 	/**
 	 * Should we be preventing friendly fire?
@@ -442,20 +428,6 @@ public class CombatUtil {
 		if (defenderTB != null && attackerTB != null && defenderTB.getType() == TownBlockType.ARENA && attackerTB.getType() == TownBlockType.ARENA)
 			return true;
 		return false;
-	}
-	
-	/**
-	 * @deprecated as of 0.96.7.1. Use {@link CombatUtil#isArenaPlot(Player, Player)}.
-	 * Return true if both attacker and defender are in Arena Plots.
-	 * 
-	 * @param attacker - Attacking Player
-	 * @param defender - Defending Player (receiving damage)
-	 * @return true if both players in an Arena plot.
-	 */
-	@Deprecated
-	public static boolean isPvPPlot(Player attacker, Player defender) {
-
-		return isArenaPlot(attacker, defender);
 	}
 
 	/**
